move options to util
