+++ snapdragon
@@ -1,12 +1,12 @@
 #!/usr/bin/env ruby
 
 require_relative 'src/colour_string.rb'
-require_relative 'src/options.rb'
 require_relative 'src/tokenizer/lexer.rb'
 require_relative 'src/tokenizer/reader.rb'
 require_relative 'src/util/logger.rb'
+require_relative 'src/util/options.rb'
 
-options = read_options
+options = Util::Options.parse_arguments
 
 Util::Logger.setup options
